+++ Changelog.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.13.6] - 2022-10-08
+
 * Fix `maturin develop` in Windows conda virtual environment in [#1146](https://github.com/PyO3/maturin/pull/1146)
 * Fix build for crate using `pyo3` and `build.rs` without `cdylib` crate type in [#1150](https://github.com/PyO3/maturin/pull/1150)
 * Fix build on some 32-bit platform by downgrading `indicatif` in [#1163](https://github.com/PyO3/maturin/pull/1163)
@@ -715,7 +717,8 @@ points-0.1.0-py2.py3-none-manylinux1_x86_64.whl | 2,8M | 752K | 85K
 
  * Initial Release
 
-[Unreleased]: https://github.com/pyo3/maturin/compare/v0.13.5...HEAD
+[Unreleased]: https://github.com/pyo3/maturin/compare/v0.13.6...HEAD
+[0.13.6]: https://github.com/pyo3/maturin/compare/v0.13.5...v0.13.6
 [0.13.5]: https://github.com/pyo3/maturin/compare/v0.13.4...v0.13.5
 [0.13.4]: https://github.com/pyo3/maturin/compare/v0.13.3...v0.13.4
 [0.13.3]: https://github.com/pyo3/maturin/compare/v0.13.2...v0.13.3